requirements.txt
aws-cdk-lib==2.219.0
constructs>=10.0.0,<11.0.0
boto3>=1.40.0
awswrangler>=3.13.0
requests>=2.32.0
python-json-logger>=4.0.0
